requirements.txt
torch
PySoundFile
torchaudio
pandas
numpy
tqdm
pyyaml
speechbrain
matplotlib
scikit-learn
tensorflow
scipy
wandb
